collectable (not finished)
placed a collectable, removes when player touches and give more health

+++ CS4700_FinalProject/Assets/Scripts/Player/PlayerHealth.cs
@@ -46,7 +46,7 @@ public class PlayerHealth : MonoBehaviour
         }
     }
 
-    public void IncreaseHealth(int value)
+    public void IncreaseHealth(float value)
     {
         if(health + value >= maxHealth) // don't allow player's health to exceed max
         {
